+++ psgscoring/__init__.py
@@ -82,7 +82,7 @@ from .utils import (
     safe_r,
 )
 
-__version__ = "0.5.0"
+__version__ = "0.5.1"
 __all__ = [
     # Master
     "run_pneumo_analysis",